requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
chardet==4.0.0
click==8.1.7
dataclasses==0.6
deprecation==2.1.0
fastapi==0.103.1
gotrue==1.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
postgrest==0.10.8
postgrest-py==0.4.0
py==1.11.0
pydantic==2.3.0
pydantic_core==2.6.3
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
toml==0.10.2
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
websockets==10.4
